refactor: Vote 도메인 관련 컴포넌트 한 군데에 모음
