+++ frontend/src/components/vote/Vote/Vote.styles.tsx
