+++ frontend/src/components/vote/VoteGenerateButton/VoteGenerateButton.tsx
