+++ frontend/src/components/vote/VoteItem/VoteItem.styles.tsx
